move utility.hpp to another directory, refactoring
