+++ ngraph/frontend/paddlepaddle/src/op/matmul.cpp
@@ -16,7 +16,7 @@
 
 #include <ngraph/opsets/opset6.hpp>
 #include "matmul.hpp"
-#include "utility.hpp"
+#include <paddlepaddle_frontend/utility.hpp>
 
 namespace ngraph {
 namespace frontend {
